A few bits and pieces being cleared up.
Including reversing an unwise change I made to page_service
yesterday.
