--- application/cms/utils.py
@@ -1,26 +1,0 @@
-import json
-
-from datetime import date
-from functools import wraps
-
-from flask import render_template
-from flask_security import current_user
-
-
-def internal_user_required(f):
-    @wraps(f)
-    def decorated_function(*args, **kwargs):
-        if current_user.is_anonymous or current_user.is_internal_user():
-            return f(*args, **kwargs)
-        else:
-            return render_template('static_site/not_allowed.html')
-    return decorated_function
-
-
-class DateEncoder(json.JSONEncoder):
-
-    def default(self, o):
-        if isinstance(o, date):
-            return o.isoformat()
-
-        return json.JSONEncoder.default(self, o)
