+++ application/sitebuilder/build.py
@@ -33,8 +33,8 @@ def do_it(application):
             build_measure_pages(page_service, subtopics, topic, topic_dir, beta_publication_states)
 
         build_other_static_pages(build_dir)
-        push_site(build_dir, build_timestamp)
-        clear_up(build_dir)
+        # push_site(build_dir, build_timestamp)
+        # clear_up(build_dir)
 
 
 def build_subtopic_pages(subtopics, topic, topic_dir):
