+++ application/templates/error/401.html
@@ -1,8 +1,8 @@
 {% extends "base.html" %}
 {% block title %}Unauthorized{% endblock %}
 {% block content %}
-    <div class="row">
-        <div class="large-offset-2 large-6 columns">
+    <div class='grid-row'>
+        <div class='column-full'>
             <h1>401</h1>
             <p>You are not authorized to see this page.</p>
         </div>
