+++ application/templates/error/404.html
@@ -1,8 +1,8 @@
 {% extends "base.html" %}
 {% block title %}Page not found{% endblock %}
 {% block content %}
-    <div class="row">
-        <div class="large-offset-2 large-6 columns">
+    <div class='grid-row'>
+        <div class='column-full'>
             <h1>404</h1>
             <p>Sorry, that page doesn't exist.</p>
         </div>
