+++ application/templates/error/500.html
@@ -1,8 +1,8 @@
 {% extends "base.html" %}
 {% block title %}Server error{% endblock %}
 {% block content %}
-    <div class="row">
-        <div class="large-offset-2 large-6 columns">
+    <div class='grid-row'>
+        <div class='column-full'>
             <h1>500</h1>
             <p>Sorry, something went wrong on our system.</p>
         </div>
