+++ application/utils.py
@@ -1,5 +1,11 @@
+import json
 import sys
 import logging
+from datetime import date
+from functools import wraps
+
+from flask import render_template
+from flask_login import current_user
 
 
 def setup_module_logging(logger, level):
@@ -10,3 +16,30 @@ def setup_module_logging(logger, level):
     logger.addHandler(handler)
     logger.setLevel(level)
     return logger
+
+
+def get_bool(param):
+    if str(param).lower().strip() in ['true', 't', 'yes', 'y', 'on', '1']:
+        return True
+    elif str(param).lower().strip() in ['false', 'f', 'no', 'n', 'off', '0']:
+        return False
+    return False
+
+
+def internal_user_required(f):
+    @wraps(f)
+    def decorated_function(*args, **kwargs):
+        if current_user.is_anonymous or current_user.is_internal_user():
+            return f(*args, **kwargs)
+        else:
+            return render_template('static_site/not_allowed.html')
+    return decorated_function
+
+
+class DateEncoder(json.JSONEncoder):
+
+    def default(self, o):
+        if isinstance(o, date):
+            return o.isoformat()
+
+        return json.JSONEncoder.default(self, o)
